Preprocessing corrections done, subdivided main class, made further justifications for baseline, added consistent docstring and function definition format. Pending stack implementation and further error handling...
